// packages/shared/src/lib/tables/cell.d.ts
import { Static } from '@sinclair/typebox';
export declare const Cell: import("@sinclair/typebox").TObject<{
    recordId: import("@sinclair/typebox").TString;
    fieldId: import("@sinclair/typebox").TString;
    projectId: import("@sinclair/typebox").TString;
    value: import("@sinclair/typebox").TUnknown;
    id: import("@sinclair/typebox").TString;
    created: import("@sinclair/typebox").TString;
    updated: import("@sinclair/typebox").TString;
}>;
export type Cell = Static<typeof Cell>;